controlling force powers and added a buy time
allow holocrons outside of the gametype; add a door open/close function for the end/start of a round to a buy phase; modify the func_door and trigger_teleport to function with the open/close functions even if gtk doesnt directly support it

+++ code/game/ai_main.h
@@ -234,7 +234,7 @@ typedef struct bot_state_s
 
 	float				destinationGrabTime;
 	float				wpSeenTime;
-	int					wpTravelTime;
+	float				wpTravelTime; // PowTecH - Basically an int at this point but scared to change it
 	float				wpDestSwitchTime;
 	float				wpSwitchTime;
 	float				wpDestIgnoreTime;
